+++ src/Chartjs/Twig/ChartExtension.php
@@ -12,6 +12,7 @@
 namespace Symfony\UX\Chartjs\Twig;
 
 use Symfony\UX\Chartjs\Model\Chart;
+use Symfony\WebpackEncoreBundle\Twig\StimulusTwigExtension;
 use Twig\Environment;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFunction;
@@ -24,6 +25,13 @@ use Twig\TwigFunction;
  */
 class ChartExtension extends AbstractExtension
 {
+    private $stimulus;
+
+    public function __construct(StimulusTwigExtension $stimulus)
+    {
+        $this->stimulus = $stimulus;
+    }
+
     public function getFunctions(): array
     {
         return [
@@ -35,11 +43,13 @@ class ChartExtension extends AbstractExtension
     {
         $chart->setAttributes(array_merge($chart->getAttributes(), $attributes));
 
-        $html = '
-            <canvas
-                data-controller="'.trim($chart->getDataController().' symfony--ux-chartjs--chart').'"
-                data-view="'.twig_escape_filter($env, json_encode($chart->createView()), 'html_attr').'"
-        ';
+        $controllers = [];
+        if ($chart->getDataController()) {
+            $controllers[$chart->getDataController()] = [];
+        }
+        $controllers['@symfony/ux-chartjs/chart'] = ['view' => $chart->createView()];
+
+        $html = '<canvas '.$this->stimulus->renderStimulusController($env, $controllers).' ';
 
         foreach ($chart->getAttributes() as $name => $value) {
             if ('data-controller' === $name) {