+++ src/LazyImage/Tests/Kernel/AppKernelTrait.php
@@ -18,12 +18,12 @@ namespace Symfony\UX\LazyImage\Tests\Kernel;
  */
 trait AppKernelTrait
 {
-    public function getCacheDir()
+    public function getCacheDir(): string
     {
         return $this->createTmpDir('cache');
     }
 
-    public function getLogDir()
+    public function getLogDir(): string
     {
         return $this->createTmpDir('logs');
     }