+++ src/Cropperjs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## 2.0
+
+-   Support for `stimulus` version 2 was removed and support for `@hotwired/stimulus`
+    version 3 was added. See the [@symfony/stimulus-bridge CHANGELOG](https://github.com/symfony/stimulus-bridge/blob/main/CHANGELOG.md#300)
+    for more details.
+-   The individual Cropper.js options in `CropperType` were moved under
+    a single `cropper_options` option.
+-   Support added for Symfony 6
+
 ## 1.3
 
 -   [DEPENDENCY CHANGE] `cropperjs` is no longer included automatically (#93)